+++ tests/test_bishop.py
@@ -5,22 +5,18 @@ from game.pieces.pawn import Pawn
 
 class TestBishop(unittest.TestCase):
 
-    def setUp(self):
-        self.board = Board()  # Esto lo hice para que las piezas con el tablero inicial 
-        for row in range(8):  # no me molesten al momento de hacer las pruebas de movimientos
-            for col in range(8):
-                self.board.set_piece(row, col, None)
-
     def test_str(self):
-        bishop = Bishop("WHITE", self.board)
+        board = Board(for_test=True)
+        bishop = Bishop("WHITE", board)
         self.assertEqual(
             str(bishop),
             "♝",
         )
 
     def test_move_diagonal_up_left(self):
-        bishop = Bishop("WHITE", self.board)
-        self.board.set_piece(4, 4, bishop)
+        board = Board(for_test=True)
+        bishop = Bishop("WHITE", board)
+        board.set_piece(4, 4, bishop)
         possibles = bishop.possible_positions_diagonal_up_left(4, 4)
         self.assertEqual(
             possibles,
@@ -28,9 +24,10 @@ class TestBishop(unittest.TestCase):
         )
 
     def test_move_diagonal_up_left_with_obstacle(self):
-        bishop = Bishop("WHITE", self.board)
-        self.board.set_piece(4, 4, bishop)
-        self.board.set_piece(6, 2, Pawn("WHITE", self.board))
+        board = Board(for_test=True)
+        bishop = Bishop("WHITE", board)
+        board.set_piece(4, 4, bishop)
+        board.set_piece(6, 2, Pawn("WHITE", board))
         possibles = bishop.possible_positions_diagonal_up_left(4, 4)
         self.assertEqual(
             possibles,
@@ -38,9 +35,10 @@ class TestBishop(unittest.TestCase):
         )
 
     def test_move_diagonal_up_left_attack(self):
-        bishop = Bishop("WHITE", self.board)
-        self.board.set_piece(4, 4, bishop)
-        self.board.set_piece(6, 2, Pawn("BLACK", self.board))
+        board = Board(for_test=True)
+        bishop = Bishop("WHITE", board)
+        board.set_piece(4, 4, bishop)
+        board.set_piece(6, 2, Pawn("BLACK", board))
         possibles = bishop.possible_positions_diagonal_up_left(4, 4)
         self.assertEqual(
             possibles,
@@ -48,8 +46,9 @@ class TestBishop(unittest.TestCase):
         )
 
     def test_move_diagonal_up_right(self):
-        bishop = Bishop("WHITE", self.board)
-        self.board.set_piece(4, 4, bishop)
+        board = Board(for_test=True)
+        bishop = Bishop("WHITE", board)
+        board.set_piece(4, 4, bishop)
         possibles = bishop.possible_positions_diagonal_up_right(4, 4)
         self.assertEqual(
             possibles,
@@ -57,9 +56,10 @@ class TestBishop(unittest.TestCase):
         )
 
     def test_move_diagonal_up_right_with_obstacle(self):
-        bishop = Bishop("WHITE", self.board)
-        self.board.set_piece(4, 4, bishop)
-        self.board.set_piece(6, 6, Pawn("WHITE", self.board))
+        board = Board(for_test=True)
+        bishop = Bishop("WHITE", board)
+        board.set_piece(4, 4, bishop)
+        board.set_piece(6, 6, Pawn("WHITE", board))
         possibles = bishop.possible_positions_diagonal_up_right(4, 4)
         self.assertEqual(
             possibles,
@@ -67,9 +67,10 @@ class TestBishop(unittest.TestCase):
         )
 
     def test_move_diagonal_up_right_attack(self):
-        bishop = Bishop("WHITE", self.board)
-        self.board.set_piece(4, 4, bishop)
-        self.board.set_piece(6, 6, Pawn("BLACK", self.board))
+        board = Board(for_test=True)
+        bishop = Bishop("WHITE", board)
+        board.set_piece(4, 4, bishop)
+        board.set_piece(6, 6, Pawn("BLACK", board))
         possibles = bishop.possible_positions_diagonal_up_right(4, 4)
         self.assertEqual(
             possibles,
@@ -77,8 +78,9 @@ class TestBishop(unittest.TestCase):
         )
 
     def test_move_diagonal_down_right(self):
-        bishop = Bishop("WHITE", self.board)
-        self.board.set_piece(4, 4, bishop)
+        board = Board(for_test=True)
+        bishop = Bishop("WHITE", board)
+        board.set_piece(4, 4, bishop)
         possibles = bishop.possible_positions_diagonal_down_right(4, 4)
         self.assertEqual(
             possibles,
@@ -86,9 +88,10 @@ class TestBishop(unittest.TestCase):
         )
 
     def test_move_diagonal_down_right_with_obstacle(self):
-        bishop = Bishop("WHITE", self.board)
-        self.board.set_piece(4, 4, bishop)
-        self.board.set_piece(2, 6, Pawn("WHITE", self.board))
+        board = Board(for_test=True)
+        bishop = Bishop("WHITE", board)
+        board.set_piece(4, 4, bishop)
+        board.set_piece(2, 6, Pawn("WHITE", board))
         possibles = bishop.possible_positions_diagonal_down_right(4, 4)
         self.assertEqual(
             possibles,
@@ -96,9 +99,10 @@ class TestBishop(unittest.TestCase):
         )
 
     def test_move_diagonal_down_right_attack(self):
-        bishop = Bishop("WHITE", self.board)
-        self.board.set_piece(4, 4, bishop)
-        self.board.set_piece(2, 6, Pawn("BLACK", self.board))
+        board = Board(for_test=True)
+        bishop = Bishop("WHITE", board)
+        board.set_piece(4, 4, bishop)
+        board.set_piece(2, 6, Pawn("BLACK", board))
         possibles = bishop.possible_positions_diagonal_down_right(4, 4)
         self.assertEqual(
             possibles,
@@ -106,8 +110,9 @@ class TestBishop(unittest.TestCase):
         )
 
     def test_move_diagonal_down_left(self):
-        bishop = Bishop("WHITE", self.board)
-        self.board.set_piece(4, 4, bishop)
+        board = Board(for_test=True)
+        bishop = Bishop("WHITE", board)
+        board.set_piece(4, 4, bishop)
         possibles = bishop.possible_positions_diagonal_down_left(4, 4)
         self.assertEqual(
             possibles,
@@ -115,9 +120,10 @@ class TestBishop(unittest.TestCase):
         )
 
     def test_move_diagonal_down_left_with_obstacle(self):
-        bishop = Bishop("WHITE", self.board)
-        self.board.set_piece(4, 4, bishop)
-        self.board.set_piece(2, 2, Pawn("WHITE", self.board))
+        board = Board(for_test=True)
+        bishop = Bishop("WHITE", board)
+        board.set_piece(4, 4, bishop)
+        board.set_piece(2, 2, Pawn("WHITE", board))
         possibles = bishop.possible_positions_diagonal_down_left(4, 4)
         self.assertEqual(
             possibles,
@@ -125,9 +131,10 @@ class TestBishop(unittest.TestCase):
         )
 
     def test_move_diagonal_down_left_attack(self):
-        bishop = Bishop("WHITE", self.board)
-        self.board.set_piece(4, 4, bishop)
-        self.board.set_piece(2, 2, Pawn("BLACK", self.board))
+        board = Board(for_test=True)
+        bishop = Bishop("WHITE", board)
+        board.set_piece(4, 4, bishop)
+        board.set_piece(2, 2, Pawn("BLACK", board))
         possibles = bishop.possible_positions_diagonal_down_left(4, 4)
         self.assertEqual(
             possibles,
@@ -135,8 +142,9 @@ class TestBishop(unittest.TestCase):
         )
 
     def test_invalid_move_straight(self):
-        self.board.set_piece(4, 4, Bishop("WHITE", self.board))  
-        bishop = self.board.get_piece(4, 4)  
+        board = Board(for_test=True)
+        board.set_piece(4, 4, Bishop("WHITE", board))  
+        bishop = board.get_piece(4, 4)  
         is_possible = bishop.valid_positions(
             from_row=4,
             from_col=4,