+++ tests/test_pawn.py
@@ -5,15 +5,11 @@ from game.pieces.queen import Queen
 
 class TestPawn(unittest.TestCase):
 
-    def setUp(self):
-        self.board = Board()  
-        for row in range(8):
-            for col in range(8):
-                self.board.set_piece(row, col, None)
+   
 
     def test_str(self):
         board = Board()
-        pawn = Pawn("WHITE", self.board)
+        pawn = Pawn("WHITE", board)
         self.assertEqual(
             str(pawn),
             "♟",
@@ -22,7 +18,7 @@ class TestPawn(unittest.TestCase):
     # Blancos
 
     def test_move_white_initial(self):
-        board = Board()
+        board = Board(for_test=True)
         pawn = Pawn("WHITE", board)
         possibles = pawn.possible_movement_vertical_up_and_down(6, 0)
         self.assertEqual(
@@ -31,7 +27,7 @@ class TestPawn(unittest.TestCase):
         )
 
     def test_move_white_not_initial(self):
-        board = Board()
+        board = Board(for_test=True)
         pawn = Pawn("WHITE", board)
         possibles = pawn.possible_movement_vertical_up_and_down(4, 2)
         self.assertEqual(
@@ -40,7 +36,7 @@ class TestPawn(unittest.TestCase):
         )
 
     def test_move_white_initial_block_other_color(self):
-        board = Board()
+        board = Board(for_test=True)
         board.set_piece(4, 0, Pawn("BLACK", board))
         pawn = Pawn("WHITE", board)
         possibles = pawn.possible_movement_vertical_up_and_down(6, 0)
@@ -50,7 +46,7 @@ class TestPawn(unittest.TestCase):
         )
 
     def test_move_white_not_initial_block_other_color(self):
-        board = Board()
+        board = Board(for_test=True)
         board.set_piece(3, 2, Pawn("BLACK", board))
         pawn = Pawn("WHITE", board)
         possibles = pawn.possible_movement_vertical_up_and_down(4, 2)
@@ -61,7 +57,7 @@ class TestPawn(unittest.TestCase):
 
 
     def test_move_white_not_initial_block(self):
-        board = Board()
+        board = Board(for_test=True)
         board.set_piece(3, 2, Pawn("WHITE", board))
         pawn = Pawn("WHITE", board)
         possibles = pawn.possible_movement_vertical_up_and_down(4, 2)
@@ -71,7 +67,7 @@ class TestPawn(unittest.TestCase):
         )
 
     def test_move_white_initial_block(self):
-        board = Board()
+        board = Board(for_test=True)
         board.set_piece(4, 0, Pawn("WHITE", board))
         pawn = Pawn("WHITE", board)
         possibles = pawn.possible_movement_vertical_up_and_down(6, 0)
@@ -81,7 +77,7 @@ class TestPawn(unittest.TestCase):
         )
 
     def test_move_white_not_initial_block(self):
-        board = Board()
+        board = Board(for_test=True)
         board.set_piece(3, 2, Pawn("WHITE", board))
         pawn = Pawn("WHITE", board)
         possibles = pawn.possible_movement_vertical_up_and_down(4, 2)
@@ -91,7 +87,7 @@ class TestPawn(unittest.TestCase):
         )
 
     def test_move_white_attack_right(self):
-        board = Board()
+        board = Board(for_test=True)
         board.set_piece(5, 1, Pawn("BLACK", board))
         pawn = Pawn("WHITE", board)
         possibles = pawn.possible_attack_diagonal(6, 0)
@@ -101,7 +97,7 @@ class TestPawn(unittest.TestCase):
         )
 
     def test_move_white_attack_left(self):
-        board = Board()
+        board = Board(for_test=True)
         board.set_piece(5, 1, Pawn("BLACK", board))
         pawn = Pawn("WHITE", board)
         possibles = pawn.possible_attack_diagonal(6, 2)
@@ -111,7 +107,7 @@ class TestPawn(unittest.TestCase):
         )
 
     def test_move_white_not_attack_right(self):
-        board = Board()
+        board = Board(for_test=True)
         board.set_piece(5, 1, Pawn("WHITE", board))
         pawn = Pawn("WHITE", board)
         possibles = pawn.possible_attack_diagonal(6, 0)
@@ -121,7 +117,7 @@ class TestPawn(unittest.TestCase):
         )
 
     def test_move_white_not_attack_left(self):
-        board = Board()
+        board = Board(for_test=True)
         board.set_piece(5, 1, Pawn("WHITE", board))
         pawn = Pawn("WHITE", board)
         possibles = pawn.possible_attack_diagonal(6, 2)
@@ -133,7 +129,7 @@ class TestPawn(unittest.TestCase):
     # # Negros
 
     def test_move_black_initial(self):
-        board = Board()
+        board = Board(for_test=True)
         pawn = Pawn("BLACK", board)
         possibles = pawn.possible_movement_vertical_up_and_down(1, 0)
         self.assertEqual(
@@ -142,7 +138,7 @@ class TestPawn(unittest.TestCase):
         )
 
     def test_move_black_not_initial(self):
-        board = Board()
+        board = Board(for_test=True)
         pawn = Pawn("BLACK", board)
         possibles = pawn.possible_movement_vertical_up_and_down(2, 2)
         self.assertEqual(
@@ -151,7 +147,7 @@ class TestPawn(unittest.TestCase):
         )
 
     def test_move_black_not_initial_block(self):
-        board = Board()
+        board = Board(for_test=True)
         board.set_piece(3, 2, Pawn("BLACK", board))
         pawn = Pawn("BLACK", board)
         possibles = pawn.possible_movement_vertical_up_and_down(2, 2)
@@ -161,7 +157,7 @@ class TestPawn(unittest.TestCase):
         )
     
     def test_move_black_initial_block(self):
-        board = Board()
+        board = Board(for_test=True)
         board.set_piece(3, 0, Pawn("BLACK", board))
         pawn = Pawn("BLACK", board)
         possibles = pawn.possible_movement_vertical_up_and_down(1, 0)
@@ -171,7 +167,7 @@ class TestPawn(unittest.TestCase):
         )
 
     def test_move_black_not_initial_block_other_color(self):
-        board = Board()
+        board = Board(for_test=True)
         board.set_piece(3, 2, Pawn("WHITE", board))
         pawn = Pawn("BLACK", board)
         possibles = pawn.possible_movement_vertical_up_and_down(2, 2)
@@ -181,7 +177,7 @@ class TestPawn(unittest.TestCase):
         )
 
     def test_move_black_initial_block_other_color(self):
-        board = Board()
+        board = Board(for_test=True)
         board.set_piece(3, 0, Pawn("WHITE", board))
         pawn = Pawn("BLACK", board)
         possibles = pawn.possible_movement_vertical_up_and_down(1, 0)
@@ -191,7 +187,7 @@ class TestPawn(unittest.TestCase):
         )
 
     def test_move_black_attack_right(self):
-        board = Board()
+        board = Board(for_test=True)
         board.set_piece(3, 1, Pawn("WHITE", board))
         pawn = Pawn("BLACK", board)
         possibles = pawn.possible_attack_diagonal(2, 0)
@@ -201,7 +197,7 @@ class TestPawn(unittest.TestCase):
         )
 
     def test_move_black_attack_left(self):
-        board = Board()
+        board = Board(for_test=True)
         board.set_piece(3, 6, Pawn("WHITE", board))
         pawn = Pawn("BLACK", board)
         possibles = pawn.possible_attack_diagonal(2, 7)
@@ -211,7 +207,7 @@ class TestPawn(unittest.TestCase):
        )
     
     def test_move_black_not_attack_right(self):
-        board = Board()
+        board = Board(for_test=True)
         board.set_piece(3, 1, Pawn("BLACK", board))
         pawn = Pawn("BLACK", board)
         possibles = pawn.possible_attack_diagonal(2, 0)
@@ -221,7 +217,7 @@ class TestPawn(unittest.TestCase):
         )
     
     def test_move_black_not_attack_left(self):
-        board = Board()
+        board = Board(for_test=True)
         board.set_piece(3, 6, Pawn("BLACK", board))
         pawn = Pawn("BLACK", board)
         possibles = pawn.possible_attack_diagonal(2, 7)
@@ -231,14 +227,14 @@ class TestPawn(unittest.TestCase):
         )
 
     # def test_promote_pawn_white(self):
-    #     board = Board()
+    #     board = Board(for_test=True))
     #     board.set_piece(1, 0, Pawn("WHITE", board))
     #     pawn = Pawn("WHITE", board)
     #     board.promote_pawn(1, 0, pawn)
     #     self.assertIsInstance(board.get_piece(0, 0), Queen)
     
     # def test_promote_pawn_black(self):
-    #     board = Board()
+    #     board = Board(for_test=True))
     #     board.set_piece(6, 0, Pawn("BLACK", board))
     #     pawn = Pawn("BLACK", board)
     #     board.promote_pawn(6, 0, pawn)
@@ -246,9 +242,9 @@ class TestPawn(unittest.TestCase):
 
 
     def test_invalid_move_white(self):
-
-        self.board.set_piece(6, 1, Pawn("WHITE", self.board))  
-        pawn = self.board.get_piece(6, 1)  
+        board = Board(for_test=True)
+        board.set_piece(6, 1, Pawn("WHITE", board))  
+        pawn = board.get_piece(6, 1)  
         is_possible = pawn.valid_positions(
             from_row=6,
             from_col=1,
@@ -258,10 +254,9 @@ class TestPawn(unittest.TestCase):
         self.assertFalse(is_possible)
 
     def test_invalid_move_black(self):
-
-
-        self.board.set_piece(1, 1, Pawn("BLACK", self.board))  
-        pawn = self.board.get_piece(1, 1)  
+        board = Board(for_test=True)
+        board.set_piece(1, 1, Pawn("BLACK", board))  
+        pawn = board.get_piece(1, 1)  
         is_possible = pawn.valid_positions(
             from_row=1,
             from_col=1,